+++ daydreaming/src/main/java/com/brainydroid/daydreaming/db/SliderSubQuestion.java
@@ -15,7 +15,7 @@ public class SliderSubQuestion {
     @SuppressWarnings("FieldCanBeLocal")
     private boolean notApplyAllowed = false;
     @SuppressWarnings("FieldCanBeLocal")
-    private boolean showHints = false;
+    private boolean showLiveIndication = false;
 
     public synchronized String getText() {
         return text;
@@ -37,8 +37,8 @@ public class SliderSubQuestion {
         return notApplyAllowed;
     }
 
-    public synchronized boolean getShowHints() {
-        return showHints;
+    public synchronized boolean getShowLiveIndication() {
+        return showLiveIndication;
     }
 
 }