+++ daydreaming/src/main/java/com/brainydroid/daydreaming/db/StarRatingSubQuestion.java
@@ -19,7 +19,7 @@ public class StarRatingSubQuestion {
     @SuppressWarnings("FieldCanBeLocal")
     private boolean notApplyAllowed = false;
     @SuppressWarnings("FieldCanBeLocal")
-    private boolean showHints = false;
+    private boolean showLiveIndication = false;
 
     public synchronized String getGlossaryText() {
         return glossaryText;
@@ -49,8 +49,8 @@ public class StarRatingSubQuestion {
         return notApplyAllowed;
     }
 
-    public synchronized boolean getShowHints() {
-        return showHints;
+    public synchronized boolean getShowLiveIndication() {
+        return showLiveIndication;
     }
 
 }