+++ daydreaming/src/main/java/com/brainydroid/daydreaming/network/ProfileFactory.java
@@ -11,7 +11,7 @@ public interface ProfileFactory {
                           @Assisted("gender") String gender,
                           @Assisted("education") String education,
                           HashMap<String, Integer> tipiAnswers,
-                          @Assisted("parametersVersion") int parametersVersion,
+                          @Assisted("parametersVersion") String parametersVersion,
                           @Assisted("appVersionName") String appVersionName,
                           @Assisted("appVersionCode") int appVersionCode,
                           @Assisted("mode") String mode);