+++ daydreaming/src/main/java/com/brainydroid/daydreaming/ui/firstlaunchsequence/FirstLaunch01DescriptionActivity.java
@@ -1,7 +1,6 @@
 package com.brainydroid.daydreaming.ui.firstlaunchsequence;
 
 import android.annotation.TargetApi;
-import android.content.Intent;
 import android.graphics.Point;
 import android.os.Build;
 import android.os.Bundle;
@@ -13,14 +12,16 @@ import android.view.ViewGroup;
 import android.widget.ImageButton;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
 import com.brainydroid.daydreaming.R;
 import com.brainydroid.daydreaming.background.Logger;
 import com.brainydroid.daydreaming.db.Util;
-import roboguice.inject.ContentView;
-import roboguice.inject.InjectView;
 
 import java.io.IOException;
 import java.io.InputStream;
+
+import roboguice.inject.ContentView;
+import roboguice.inject.InjectView;
 
 /**
  * Activity at first launch
@@ -40,7 +41,7 @@ public class FirstLaunch01DescriptionActivity extends FirstLaunchActivity {
     private static String TAG = "FirstLaunch01DescriptionActivity";
     @InjectView(R.id.firstLaunchDescription_textDescription) TextView description;
     @InjectView(R.id.firstLaunchDescription_scrollview_layout)
-    LinearLayout myScrollLayout;
+    LinearLayout scrollLayout;
 
     protected ImageButton nextButton;
 
@@ -51,27 +52,24 @@ public class FirstLaunch01DescriptionActivity extends FirstLaunchActivity {
 
         nextButton = (ImageButton)findViewById(R.id.firstLaunchDescription_buttonNext);
 
-
         makeLayoutDesign();
         populateDescription();
         setButton();
 
-
         setRobotoFont(this);
-        Linkify.addLinks(description,Linkify.WEB_URLS | Linkify.EMAIL_ADDRESSES);
-
+        Linkify.addLinks(description, Linkify.WEB_URLS | Linkify.EMAIL_ADDRESSES);
     }
 
     /**
      * OnClick Listener, launches next activity in sequence when next button of view is clicked on
      */
     public void onClick_buttonNext(@SuppressWarnings("UnusedParameters") View view) {
-        Logger.d(TAG, "Next button clicked -> launching consent dialog");
+        Logger.d(TAG, "Next button clicked -> launching terms activity");
         launchNextActivity(FirstLaunch02TermsActivity.class);
     }
 
     @TargetApi(13)
-    public void makeLayoutDesign(){
+    public void makeLayoutDesign() {
         int y;
         Display display = getWindowManager().getDefaultDisplay();
 
@@ -84,12 +82,14 @@ public class FirstLaunch01DescriptionActivity extends FirstLaunchActivity {
             //noinspection deprecation
             y = display.getHeight();
         }
-        int screen_height = (int)(y / getResources().getDisplayMetrics().density);
+        int screenHeight = (int)((float)y / getResources().getDisplayMetrics().density);
 
-        int personComplexDpHeight = (int) (getResources().getDimension(R.dimen.cloud_cone_person_heigth) / getResources().getDisplayMetrics().density);     // size of cloud/person/cone complex
-        ViewGroup.LayoutParams params = myScrollLayout.getLayoutParams();
+        // Size of cloud/person/cone complex
+        int personComplexDpHeight = (int)(getResources().getDimension(R.dimen.cloud_cone_person_heigth) / getResources().getDisplayMetrics().density);
+        ViewGroup.LayoutParams params = scrollLayout.getLayoutParams();
 
-        params.height = (int)( ((float)(screen_height-personComplexDpHeight))*(getResources().getDisplayMetrics().density) )+1;
+        params.height = (int)(((float)(screenHeight - personComplexDpHeight)) *
+                               getResources().getDisplayMetrics().density) + 1;
     }
 
     /**
@@ -107,28 +107,20 @@ public class FirstLaunch01DescriptionActivity extends FirstLaunchActivity {
         }
     }
 
-
-    public void setButton(){
+    public void setButton() {
         nextButton.setVisibility(View.VISIBLE);
         nextButton.setClickable(true);
-    }
-
-    /**
-     * Terms activity separated in sequence.
-     * Exiting at this point completely leaves the app.
-     */
-    @Override
-    protected void launchNextActivity(Class activity) {
-        Intent intent = new Intent(this, activity);
-        intent.setFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
-        startActivity(intent);
-        overridePendingTransition(R.anim.push_left_in, R.anim.push_left_out);
     }
 
     // Overriding parent method
     @Override
     public boolean shouldFinishIfTipiQuestionnaireCompleted() {
         return true;
+    }
+
+    @Override
+    public void backHook() {
+        Logger.v(TAG, "Leaving default transition");
     }
 
 }